File structure updated (#11)
* File structure updated

* Fix(rotation): prevent O-piece from rotating in SrsRotation

The O tetromino (2x2 square) is rotation-invariant but was previously
being rotated, causing a visual shift. Added a type check in
SrsRotation.tryRotateCW() to skip rotation for O pieces.

* Back Button Functionality, Author tags and SplashScreen modified

* refactor(ui): move UI classes into ui package and update pom mainClass

- Moved all UI-related classes into the `tetris.ui` package for better structure
- Updated pom.xml to set `tetris.ui.SplashScreen` as the mainClass

* Bug Fix: Movements issue fixed when Paused

* adding some simple comments and moving things around

* feat(ui): add centralized UIConfigurations class for window dimensions

- Introduced UIConfigurations class to store shared UI constants
- Defined default window width and height (500x500)
- Added documentation comments for clarity and future extension

* refactor(ui): apply UIConfigurations constants to all window scenes

- Updated Scene creation to use centralized UIConfigurations for width and height
- Ensures consistent window sizing across ConfigScreen, Gameplay, HighScore, MainMenu, and SplashScreen
- Simplifies future maintenance by keeping UI dimensions in a single configuration class

* refactor(ui): apply UIConfigurations constants to all window scenes

- Updated Scene creation to use centralized UIConfigurations for width and height
- Ensures consistent window sizing across ConfigScreen, Gameplay, HighScore, MainMenu, and SplashScreen
- Simplifies future maintenance by keeping UI dimensions in a single configuration class

---------
