+++ Tetris/pom.xml
@@ -33,7 +33,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.8</version>
                 <configuration>
-                    <mainClass>JavaFX</mainClass>
+                    <mainClass>tetris.ui.SplashScreen</mainClass>
                 </configuration>
             </plugin>
         </plugins>
