--- Tetris/src/main/java/tetris/JavaFX.java
@@ -1,23 +1,0 @@
-package tetris;
-
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.layout.StackPane;
-import javafx.stage.Stage;
-
-public class JavaFX extends Application {
-    @Override
-    public void start(Stage primaryStage) {
-        Button btn = new Button("Don't Click Me");
-        btn.setOnAction(e -> System.out.println("You Fail!"));
-        StackPane root = new StackPane(btn);
-        Scene scene = new Scene(root, 300, 200);
-        primaryStage.setTitle("English Test");
-        primaryStage.setScene(scene);
-        primaryStage.show();
-    }
-    public static void main(String[] args) {
-        launch(args);
-    }
-}
