+++ Tetris/src/main/java/tetris/model/Vec.java
@@ -1,0 +1,3 @@
+package tetris.model;
+
+public record Vec(int x, int y) {}
