+++ Tetris/src/main/java/tetris/model/piece/ActivePiece.java
@@ -1,0 +1,19 @@
+package tetris.model.piece;
+
+import java.util.List;
+import tetris.model.TetrominoType;
+import tetris.model.Vec;
+
+public class ActivePiece extends Piece {
+    private Vec origin; // board coords
+
+    public ActivePiece(TetrominoType t, Vec origin) { super(t); this.origin = origin; }
+
+    public Vec origin() { return origin; }
+    public void moveBy(int dx, int dy) { origin = new Vec(origin.x() + dx, origin.y() + dy); }
+    public void rotateCW() { rotation = (rotation + 1) & 3; }
+
+    public List<Vec> worldCells() {
+        return localCells().stream().map(v -> new Vec(v.x() + origin.x(), v.y() + origin.y())).toList();
+    }
+}
